+++ src/fst/fst_put_slice.py
@@ -646,7 +646,7 @@ def _code_to_slice_Compare__all_maybe_dangling(
     if op_side_left is None:
         return fst_, None
 
-    if not (op := options.get('op')):
+    if not (op := fst.FST.get_option('op', options)):
         if is_ins:  # if is insert then a dangling operator MUST be added from a non-global `op` option which must be present, op_side_left will be either True or False in that case
             raise ValueError("insertion to Compare requires an 'op' extra operator to insert")